+++ cx-portal/src/components/overlays/NotFound/index.tsx
@@ -18,47 +18,44 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
+import NotFoundContent from 'components/pages/NotFound/NotFoundContent'
 import {
-  IdentityProvider,
-  useFetchIDPListQuery,
-} from 'features/admin/idpApiSlice'
-import { show } from 'features/control/overlay/actions'
+  Button,
+  DialogActions,
+  DialogContent,
+  DialogHeader,
+} from 'cx-portal-shared-components'
+import { closeOverlay } from 'features/control/overlay/actions'
 import { useTranslation } from 'react-i18next'
 import { useDispatch } from 'react-redux'
-import { OVERLAYS } from 'types/Constants'
-import './style.scss'
 
-export const IDPListItem = ({ idp }: { idp: IdentityProvider }) => {
+export default function NotFound() {
   const { t } = useTranslation()
   const dispatch = useDispatch()
-  const state = idp.enabled ? 'enabled' : 'disabled'
-  return (
-    <div
-      className="idp-list-item"
-      onClick={() => dispatch(show(OVERLAYS.IDP, idp.identityProviderId))}
-    >
-      <span className="category">
-        {t(
-          `content.idpmanagement.idp.category.${idp.identityProviderCategoryId}`
-        )}
-      </span>
-      <span className="name">{idp.displayName || idp.alias}</span>
-      <span className="state">{t(`global.state.${state}`)}</span>
-    </div>
-  )
-}
-
-export const IDPList = () => {
-  const { data } = useFetchIDPListQuery()
+  const close = () => dispatch(closeOverlay())
 
   return (
-    <ul className="idp-list">
-      {data &&
-        data.map((idp) => (
-          <li key={idp.identityProviderId}>
-            <IDPListItem idp={idp} />
-          </li>
-        ))}
-    </ul>
+    <>
+      <DialogHeader
+        {...{
+          title: t('pages.notfound'),
+          closeWithIcon: true,
+          onCloseWithIcon: close,
+        }}
+      />
+
+      <DialogContent>
+        <NotFoundContent />
+      </DialogContent>
+
+      <DialogActions>
+        <Button variant="outlined" onClick={close}>
+          {`${t('global.actions.cancel')}`}
+        </Button>
+        <Button variant="contained" onClick={close}>
+          {`${t('global.actions.close')}`}
+        </Button>
+      </DialogActions>
+    </>
   )
 }