+++ RELEASENOTES.md
@@ -10,7 +10,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 * Feature - Connect Notifications API
 * Feature - App Release Process Step 5 - Beta Test
 * Feature - Search functionality added in Invite Business Partner page
-* Feature - Identity Provider List
+* Feature - Identity provider list and detail view
 
 
 ### 0.5.2 (2022-08-23)